Move either, maybe and try into monads directory and move stream into streams directory
