+++ pynction/monads/try_monad.py
